requirements.txt
fastapi
uvicorn
pandas
plotly
jinja2
python-dotenv
groq
fastmcp
matplotlib
python-jose[cryptography]
fastmcp
httpx
reportlab
streamlit
slowapi
pydantic>=2.0